pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "email-validator>=2.2.0",
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.75.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "flask-socketio>=5.5.1",
    "eventlet>=0.39.1",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "numpy>=2.2.5",
    "google-generativeai>=0.8.5",
    "gitpython>=3.1.44",
    "requests>=2.32.3",
    "trafilatura>=2.0.0",
    "beautifulsoup4>=4.13.4",
    "ebooklib>=0.18",
    "markdown>=3.8",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "pypdf>=5.4.0",
    "docx>=0.2.4",
    "twilio>=9.5.2",
    "rarfile>=4.2",
]
